+++ crates/netconf-proto/src/yang_push/subscription.rs
@@ -904,3 +904,47 @@ impl<'a> XmlDeserialize<'a, DatastoreSelectionFilterObjects> for DatastoreSelect
         }
     }
 }
+
+/// Module Versioning (draft-ietf-netconf-yang-notifications-versioning)
+///
+/// The `YangPushModuleVersion` structure provides information about YANG
+/// modules:
+/// - `name`: Module name
+/// - `revision`: Module revision date (e.g., "2025-04-25")
+/// - `version`: Semantic version label
+#[derive(Default, Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "kebab-case")]
+pub struct YangPushModuleVersion {
+    /// Alias 'module-name' still supported
+    /// (draft-ietf-netconf-yang-notifications-versioning < 9)
+    #[serde(alias = "module-name")]
+    pub name: String,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub revision: Option<String>,
+
+    /// Alias 'revision-label' still supported
+    /// (draft-ietf-netconf-yang-notifications-versioning < 9)
+    #[serde(alias = "revision-label")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub version: Option<String>,
+}
+
+impl YangPushModuleVersion {
+    pub const fn new(name: String, revision: Option<String>, version: Option<String>) -> Self {
+        Self {
+            name,
+            revision,
+            version,
+        }
+    }
+    pub const fn name(&self) -> &str {
+        self.name.as_str()
+    }
+    pub fn revision(&self) -> Option<&str> {
+        self.revision.as_deref()
+    }
+    pub fn version(&self) -> Option<&str> {
+        self.version.as_deref()
+    }
+}
